--- src/main/java/com/prography/lighton/auth/application/dto/TokenDTO.java
@@ -1,7 +1,0 @@
-package com.prography.lighton.auth.application.dto;
-
-public record TokenDTO(
-        String refreshToken,
-        String accessToken
-) {
-}
