+++ src/main/java/com/prography/lighton/auth/infrastructure/oauth/apple/AppleOAuthTokenService.java
@@ -4,9 +4,9 @@ package com.prography.lighton.auth.infrastructure.oauth.apple;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.prography.lighton.auth.application.exception.IdTokenParseException;
 import com.prography.lighton.auth.application.exception.InvalidTokenException;
 import com.prography.lighton.auth.infrastructure.config.AppleOAuthProperties;
+import com.prography.lighton.auth.infrastructure.oauth.apple.exception.IdTokenParseException;
 import com.prography.lighton.auth.presentation.dto.apple.AppleIdTokenPayload;
 import com.prography.lighton.auth.presentation.dto.apple.AppleUser;
 import java.security.interfaces.ECPrivateKey;
